requirements.txt
torch==1.13.1
torchvision==0.14.1
transformers==4.33.3
datasets==2.14.5
accelerate==0.21.0
pillow
numpy<2
scipy